Rename folders to certified
