+++ templates/certified/component-details.html
@@ -1,4 +1,4 @@
-{% extends "certification/base_certification.html" %}
+{% extends "certified/base_certification.html" %}
 
 {% block title %}Ubuntu certified component: {{ component.vendor_name }} {{ component.model }}{% endblock %}
