+++ templates/certified/model-details.html
@@ -1,4 +1,4 @@
-{% extends "certification/base_certification.html" %}
+{% extends "certified/base_certification.html" %}
 
 {% block title %} {{ vendor }} {{ model }} certified with Ubuntu | Ubuntu{% endblock %}
