+++ templates/certified/search-results.html
@@ -1,4 +1,4 @@
-{% extends "certification/base_certification.html" %}
+{% extends "certified/base_certification.html" %}
 
 {% block title %}Certified hardware{% endblock %}
